# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: History Wiki API
  version: 0.1.0
  description: 历史 Wiki 的公开读者 API。空数据库行为是本契约的一等状态。
  license:
    name: Unlicensed
servers:
  - url: /
security: []
paths:
...
  /api/v1/admin/users:
    get:
      operationId: listManagedUsers
      summary: 列出保留的编辑者和管理员账号
      description: 仅管理员可用；停用账号仍返回，账号不提供物理删除接口。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 账号列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedUserListResponse" }
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createManagedUser
      summary: 创建编辑者或管理员账号
      description: 仅管理员可用；不开放公开注册或邮件找回。
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateManagedUserRequest" }
      responses:
        "201":
          description: 账号已创建
          headers:
            Location:
              schema: { type: string }
            ETag:
              description: 账号当前并发版本的强 ETag。
              schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedUser" }
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-history_wiki_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 必须与 Secure CSRF Cookie 一致，并与当前服务端 Session 绑定。
      schema: { type: string, minLength: 1 }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: JSON 请求体超过此操作允许的 2 MiB 或 20 MiB 上限
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: 请求超过约定速率
      headers:
        Retry-After:
          description: 建议重试前等待的秒数
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    UserRole:
      type: string
      enum: [editor, administrator]
...
    ManagedUser:
      type: object
      additionalProperties: false
      required: [id, email, role, disabledAt, lockVersion, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/UserRole" }
        disabledAt:
          type: [string, "null"]
          format: date-time
        lockVersion: { type: integer, format: int64, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ManagedUserListResponse:
      type: object
      additionalProperties: false
      required: [users]
      properties:
        users:
          type: array
          items: { $ref: "#/components/schemas/ManagedUser" }
    CreateManagedUserRequest:
      type: object
      additionalProperties: false
      required: [email, password, role]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 1024 }
        role: { $ref: "#/components/schemas/UserRole" }
...
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }
...